print the first 50 fibonacci number

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
 
 int main(void)
 {
-int num1, num2, counter, sum;
+long int num1, num2, counter, sum;
 num1 = 0, num2 = 1;
 counter = 0;
 while (counter <= 49)
@@ -15,9 +15,9 @@ sum = num1 + num2;
 num1 = num2;
 num2 = sum;
 if (counter != 49)
-	printf("%d, ", sum);
+	printf("%ld, ", sum);
 else
-	printf("%d\n", sum);
+	printf("%ld\n", sum);
 counter++;
 }
 return (0);
